Extract XfccField and XfccHeaderParser into shared core module
Add java-buildpack-client-certificate-mapper-core module containing
XfccField (public enum) and XfccHeaderParser (public final class) with
no servlet dependency. Both jakarta and javax modules depend on core and
import from org.cloudfoundry.router. Duplicate XfccField, XfccHeaderParser
and XfccHeaderParserTest removed from both servlet modules.
