+++ pom.xml
@@ -54,6 +54,7 @@
     </scm>
 
     <modules>
+        <module>java-buildpack-client-certificate-mapper-core</module>
         <module>java-buildpack-client-certificate-mapper</module>
         <module>java-buildpack-client-certificate-mapper-jakarta</module>
         <module>java-buildpack-client-certificate-mapper-javax</module>
